+++ sample/src/samples/button/binding.js
@@ -1,0 +1,3 @@
+export class Binding {
+  isEnabled = false;
+}