+++ sample/src/samples/grid/advanced.js
@@ -1,15 +1,22 @@
 export class Advanced {
   datasource = {
     data: [{
-      ContactName: 'jeroen'
+      ContactName: 'jeroen',
+      PhoneNumbers: ['12345678', '87654321']
     }, {
-      ContactName: 'charles'
+      ContactName: 'charles',
+      PhoneNumbers: ['12345678', '87654321']
     }, {
-      ContactName: 'nikolaj'
+      ContactName: 'nikolaj',
+      PhoneNumbers: ['12345678', '87654321']
     }]
   }
 
   showAlert(message) {
     alert(message);
+  }
+
+  log(obj) {
+    console.log(obj); // eslint-disable-line no-console
   }
 }