+++ sample/src/samples/grid/custom-grid-width.js
@@ -1,0 +1,9 @@
+export class CustomGridWidth {
+  datasource = {
+    type: 'odata',
+    transport: {
+      read: '//demos.telerik.com/kendo-ui/service/Northwind.svc/Customers'
+    },
+    pageSize: 20
+  };
+}