+++ sample/src/samples/timepicker/range-selection.js
@@ -1,29 +1,34 @@
 import {bindable} from 'aurelia-templating';
+import {inject, TaskQueue} from 'aurelia-framework';
 
+@inject(TaskQueue)
 export class RangeSelection {
     @bindable startTime = '8:00 AM';
 
+    constructor(taskQueue) {
+      this.taskQueue = taskQueue;
+    }
+
     attached() {
-      this.startTimePicker = this.start.kWidget;
-      this.endTimePicker = this.end.kWidget;
+      this.taskQueue.queueTask(() => {
+        this.start.min('8:00 AM');
+        this.start.max('6:00 PM');
 
-      this.startTimePicker.min('8:00 AM');
-      this.startTimePicker.max('6:00 PM');
-
-      this.endTimePicker.min('8:00 AM');
-      this.endTimePicker.max('7:30 AM');
+        this.end.min('8:00 AM');
+        this.end.max('7:30 AM');
+      });
     }
 
     startTimeChanged() {
       if (this.startTime) {
         let time = new Date(this.startTime);
 
-        this.endTimePicker.max(time);
+        this.end.max(time);
 
-        time.setMinutes(time.getMinutes() + this.endTimePicker.options.interval);
+        time.setMinutes(time.getMinutes() + this.end.options.interval);
 
-        this.endTimePicker.min(time);
-        this.endTimePicker.value(time);
+        this.end.min(time);
+        this.end.value(time);
       }
     }
 }