+++ src/calendar/calendar.js
@@ -1,5 +1,5 @@
 import {inject} from 'aurelia-dependency-injection';
-import {customElement, bindable} from 'aurelia-templating';
+import {customElement} from 'aurelia-templating';
 import {WidgetBase} from '../common/widget-base';
 import {generateBindables} from '../common/decorators';
 import {constants} from '../common/constants';
@@ -9,8 +9,6 @@ import 'kendo.calendar.min';
 @generateBindables('kendoCalendar')
 @inject(Element, WidgetBase)
 export class Calendar {
-
-  @bindable kOptions = {};
 
   constructor(element, widgetBase, viewResources) {
     this.element = element;