+++ src/datetimepicker/datetimepicker.js
@@ -10,14 +10,17 @@ import 'kendo.datetimepicker.min';
 @inject(Element, WidgetBase)
 export class DateTimePicker {
 
-  @bindable kOptions = {};
+  @bindable kEnabled;
+  @bindable kReadOnly;
 
   constructor(element, widgetBase) {
     this.element = element;
     this.widgetBase = widgetBase
                         .control('kendoDateTimePicker')
                         .linkViewModel(this)
-                        .useValueBinding();
+                        .useValueBinding()
+                        .bindToKendo('kEnabled', 'enable')
+                        .bindToKendo('kReadOnly', 'readonly');
   }
 
   bind(ctx) {