+++ src/dropdownlist/dropdownlist.js
@@ -11,8 +11,9 @@ import 'kendo.virtuallist.min';
 @inject(Element, WidgetBase, ViewResources)
 export class DropDownList {
 
-  @bindable kOptions = {};
   @bindable kNoValueBinding = false;
+  @bindable kEnabled;
+  @bindable kReadOnly;
   @children(`${constants.elementPrefix}template`) templates;
 
   constructor(element, widgetBase, viewResources) {
@@ -20,8 +21,9 @@ export class DropDownList {
     this.widgetBase = widgetBase
                         .control('kendoDropDownList')
                         .linkViewModel(this)
-                        .useValueBinding()
-                        .useViewResources(viewResources);
+                        .useViewResources(viewResources)
+                        .bindToKendo('kEnabled', 'enable')
+                        .bindToKendo('kReadOnly', 'readonly');
   }
 
   bind(ctx) {
@@ -29,11 +31,11 @@ export class DropDownList {
   }
 
   attached() {
-    this.recreate();
-
-    if (this.kNoValueBinding) {
-      this.widgetBase.withValueBinding = false;
+    if (!this.kNoValueBinding) {
+      this.widgetBase.useValueBinding();
     }
+
+    this.recreate();
   }
 
   recreate() {