+++ src/slider/slider.js
@@ -10,13 +10,14 @@ import 'kendo.slider.min';
 @inject(Element, WidgetBase)
 export class Slider {
 
-  @bindable kOptions = {};
+  @bindable kEnabled;
 
   constructor(element, widgetBase) {
     this.element = element;
     this.widgetBase = widgetBase
                     .control('kendoSlider')
                     .linkViewModel(this)
+                    .bindToKendo('kEnabled', 'enable')
                     .useValueBinding();
   }