+++ doc/CHANGELOG.md
@@ -1,3 +1,56 @@
+<a name="0.9.0"></a>
+# [0.9.0](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/compare/0.8.1...v0.9.0) (2016-05-05)
+
+
+### Bug Fixes
+
+* **all:** fix k-enable/k-readonly binding for button, buttongroup and progressbar ([d642266](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/d642266))
+* **recreate:** destroy widgets before creating them again ([47b40ec](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/47b40ec))
+* **widget-base:** use kendo.jQuery and not jQuery ([4a9391f](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/4a9391f))
+
+### Features
+
+* **all:** support k-enabled and k-read-only binding ([401b4ef](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/401b4ef))
+* **template-compiler:** support $this in templates ([9eb7908](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/9eb7908))
+
+
+### BREAKING CHANGES
+
+* template-compiler: $$item has been removed and $this should be used now
+
+closes https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/issues/339
+
+
+
+<a name="0.8.1"></a>
+## [0.8.1](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/compare/0.8.0...v0.8.1) (2016-04-28)
+
+
+### Bug Fixes
+
+* **dependencies:** remove jquery version lock ([6e1203a](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/6e1203a))
+* **template-gatherer:** throw error when templating support is not enabled ([6412fa2](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/6412fa2))
+* **value-binding:** don't update view-model with initial value in case of empty string value ([793671b](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/793671b))
+
+
+
+<a name="0.8.0"></a>
+# [0.8.0](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/compare/0.7.1...v0.8.0) (2016-04-23)
+
+
+### Bug Fixes
+
+* **all:** allow repeaters in ak-templates ([2ffd035](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/2ffd035))
+* **index:** import minified kendo.data and improve import of repeat strategy ([f92cef2](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/f92cef2))
+* **jquery:** remove jquery version lock ([abbe694](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/abbe694))
+
+### Features
+
+* **catalog:** "test implementation" of RUN button ([aea9524](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/aea9524))
+* **catalog:** add two tutorials ([f394a54](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/commit/f394a54))
+
+
+
 <a name="0.7.1"></a>
 ## [0.7.1](https://github.com/aurelia-ui-toolkits/aurelia-kendoui-bridge/compare/0.7.0...v0.7.1) (2016-04-09)